Deleted AllProducts and moved code to Products
